% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_universe.R
\name{sim_individual_MCED}
\alias{sim_individual_MCED}
\title{Simulate an individual with and without Multicancer Early Detection (MCED) screening}
\usage{
sim_individual_MCED(
  ID,
  cancer_sites,
  rates_list,
  test_performance,
  MCED_specificity,
  other_cause_death_dist,
  starting_age,
  num_screens,
  screen_interval,
  end_time,
  sex,
  surv_param_table,
  optimistic_surv_param_table = NULL
)
}
\arguments{
\item{ID}{Numeric ID for the individual.}

\item{cancer_sites}{Vector of cancer sites (allowable values include "Anus", ” Bladder",”Breast",  ”Colorectal",
"Esophagus"  , ”Gastric", "Headandneck", "Liver", "Lung", "Pancreas", "Prostate", "Renal", "Uterine","Ovary")}

\item{rates_list}{List of transition rate matrices for each cancer site.}

\item{test_performance}{A list with early_sens, late_sens, and specificities for the test.}

\item{MCED_specificity}{Overall specificity for the MCED test (not specific to cancer site)}

\item{other_cause_death_dist}{A table representing other-cause mortality.}

\item{starting_age}{Starting age for simulation.}

\item{num_screens}{Number of screening rounds.}

\item{screen_interval}{Time interval between screenings.}

\item{end_time}{Ending time (age) of simulation.}

\item{sex}{"Male" or "Female".}

\item{surv_param_table}{Survival parameters table (for simulating cancer death).}

\item{optimistic_surv_param_table}{Survival optimistic parameters table (for simulating cancer death).}
}
\value{
A data frame with the individual's first-cancer outcomes (onset, screen/clinical diagnosis times and stages),
cancer death times with/without screening, and other-cause death.
}
\description{
This function simulates cancer outcomes for a single individual in two parallel scenarios:
with scheduled MCED screening and without screening. It generates cancer onset, screen detection,
clinical diagnosis, cancer death, and other-cause death, retaining the first cancer by earliest onset.
}
\details{
Natural history models for each cancer site are specified based on a list of transition rate matrices for each site.
The function tracks only the first cancer diagnosis based on pre-clinical onset.
Cancer-specific mortality in the screen arm assumes a stage-shift benefit of screening. That is, individuals
who are diagnosed in early stage under screening but who would have been diagnosed in late stage clinically, are assume to remain in early stage post lead-time.
For both screened and unscreened individuals, cancer mortality is projected from the point of clinical diagnosis (i.e., post lead-time) to prevent
lead-time bias.
}
\examples{
library(MCEDsim)
#Load the prefitted natural history models
data("combined_fits")
#Load the prefitted cause-specific survival models
data("parametric_surv_fits")

#' # Cancer sites in the MCED test
cancer_sites_vec <- c("Anus", "Bladder", "Breast", "Esophagus",
                     "Gastric", "Headandneck", "Liver",
                     "Lung", "Pancreas", "Prostate", "Renal", "Uterine", "Ovary")

result <- sim_individual_MCED(
  ID = 1,
  cancer_sites = cancer_sites_vec,
  rates_list = example_rates_list,
  test_performance = example_test_perf,
  other_cause_death_dist = example_death_dist,
  starting_age = 50,
  num_screens = 5,
  screen_interval = 1,
  end_time = 90,
  sex = "Male",
  surv_param_table = param_table
)
}
