% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_CRC_combine.R
\name{sim_MCED_parallel_universe_before_CRC}
\alias{sim_MCED_parallel_universe_before_CRC}
\title{Simulate a cohort of individuals with and without Multicancer Early Detection (MCED) screening}
\usage{
sim_MCED_parallel_universe_before_CRC(
  cancer_sites,
  LMST_vec,
  OMST_vec,
  test_performance_dataframe,
  MCED_specificity,
  starting_age,
  ending_age,
  num_screens,
  screen_interval,
  num_males,
  num_females,
  all_rates_male,
  all_rates_female,
  all_meta_data_female,
  all_meta_data_male,
  cdc_data,
  hmd_data,
  MCED_cdc,
  surv_param_table,
  optimistic_surv_param_table = NULL,
  simulation_seed
)
}
\arguments{
\item{cancer_sites}{Vector of cancer sites (allowable values include
"Anus",  "Bladder", "Esophagus", "Gastric", "Headandneck",
"Liver", "Lung", "Lymphoma", "Ovary", "Pancreas", "Renal", "Uterine")}

\item{LMST_vec}{Numeric vector of late mean sojourn times (years) for each cancer site.}

\item{OMST_vec}{Numeric vector of overall mean sojourn times (years) for each cancer site.}

\item{test_performance_dataframe}{Data frame with test sensitivity/specificity info.}

\item{MCED_specificity}{Overall specificity for the MCED test (not specific to cancer site)}

\item{starting_age}{Numeric starting age for simulation}

\item{ending_age}{Numeric ending age for simulation}

\item{num_screens}{Number of screening rounds.}

\item{screen_interval}{Interval between screening rounds.}

\item{num_males}{Number of male individuals to simulate.}

\item{num_females}{Number of female individuals to simulate.}

\item{all_rates_male}{List of transition matrices for males.}

\item{all_rates_female}{List of transition matrices for females.}

\item{all_meta_data_female}{Metadata for female cancer sites.}

\item{all_meta_data_male}{Metadata for male cancer sites.}

\item{cdc_data}{CDC mortality data.}

\item{hmd_data}{Human Mortality Database data.}

\item{MCED_cdc}{CDC data for MCED.}

\item{surv_param_table}{Data frame of cancer-specific survival parameters.}

\item{simulation_seed}{Set seed for each simulation}

\item{CRC_data}{CRC data.}
}
\value{
A data frame with combined simulated results for all individuals.
The function returns each individual's first cancer site,  age and stage of clinical diagnosis in
absence of screening, age and stage at screen diagnosis, the time of other-cause mortality, and the time of cancer-specific mortality in absence
and presence of screening.   Cancer diagnosis and death times are presented both with and without competing other-cause mortality.
}
\description{
This function simulates cancer outcomes in a population with a designated starting age using a "parallel universe" approach.
That is, cancer outcomes in the population are simulated with and without MCED screening.  The natural history (i.e., the times of cancer onset and clinical diagnosis) for
each individual is the same in both screening and no-screening scenarios.

The user specifies cancer sites in the MCED screening test.
The user also provides sensitivity of the tests for early and late-stage disease, where early refers
to AJCC 7 stages I-II and late, III-IV, except for pancreas cancer, where early is stage I and late, II-IV.

Natural history models are based on built-in fitted models that are calibrated to SEER 2015-2021 data by age and sex and can be specified based on
user-provided inputs about the overall mean sojourn time (OMST) and the late mean sojourn time (LMST) for each cancer site.
The function tracks only the first cancer diagnosis based on pre-clinical onset.

Other-cause mortality is based on all cause mortality tables from the Human Mortality Database that have been adjusted to remove the mortality due to
cancers included in the MCED tests.  Cancer-specific mortality in the screen arm assumes a stage-shift benefit of screening. That is, individuals
who are diagnosed in early stage under screening but who would have been diagnosed in late stage clinically, are assume to remain in early stage post lead-time.
For both screened and unscreened individuals, cancer mortality is projected from the point of clinical diagnosis (i.e., post lead-time) to prevent
lead-time bias.
}
\examples{
library(MCEDsimCarolyn)

# Load the other-cause mortality tables
data("cdc_hmd_data")
# Load the prefitted natural history models
data("combined_fits")
# Load the prefitted cause-specific survival models
data("parametric_surv_fits")
#load("/home/groups/CEDAR/MCED_sim/parametric_surv_fits.rda")

theseed      <- 1
scenario_no  <- 3

cancer_sites_vec <- c(
 "Anus",  "Bladder", "Esophagus", "Gastric", "Headandneck",
 "Liver", "Lung", "Lymphoma", "Ovary", "Pancreas", "Renal", "Uterine")

OMST_vec <- rep(2, 12)
LMST_vec <- rep(0.5, 12)

early_sens <- c(
 0.5,  0.18, 0.48, 0.33, 0.72, 0.81,
 0.40, 0.61, 0.60, 0.61, 0.07, 0.18)

late_sens <- c(
 1.00, 0.83, 0.97, 0.94, 0.93, 1.00,
 0.93, 0.94, 0.90, 0.94, 0.45, 0.81)

test_performance_dataframe <- data.frame(early_sens  = early_sens,
                                        late_sens   = late_sens,
                                        cancer_site = cancer_sites_vec)

set.seed(123)
results <- sim_MCED_parallel_universe_before_CRC(cancer_sites          = cancer_sites_vec,
                                                LMST_vec              = LMST_vec,
                                                OMST_vec              = OMST_vec,
                                                test_performance_dataframe = test_performance_dataframe,
                                                starting_age          = 45,
                                                ending_age            = 500,
                                                num_screens           = 30,
                                                screen_interval       = 1,
                                                num_males             = 50,
                                                num_females           = 50,
                                                all_rates_male        = all_rates_male,
                                                all_rates_female      = all_rates_female,
                                                all_meta_data_female  = all_meta_data_female,
                                                all_meta_data_male    = all_meta_data_male,
                                                cdc_data              = all_cause_cdc,
                                                hmd_data              = hmd_data,
                                                MCED_cdc              = MCED_cdc,
                                                surv_param_table      = param_table,
                                                MCED_specificity      = 0.995,
                                                simulation_seed       = theseed)
}
