% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicancer_sim.R
\name{sim_multiple_cancer_indiv}
\alias{sim_multiple_cancer_indiv}
\title{Simulate multiple cancer sites for an individual}
\usage{
sim_multiple_cancer_indiv(
  ID,
  rate_matrices,
  early_sensitivities,
  late_sensitivities,
  specificities,
  start.time,
  end.time,
  obs.times,
  start.states,
  cancer_sites
)
}
\arguments{
\item{ID}{Numeric identifier for the individual.}

\item{rate_matrices}{List of rate matrices corresponding to each cancer site.}

\item{early_sensitivities}{Numeric vector of early sensitivities for each cancer site.}

\item{late_sensitivities}{Numeric vector of late sensitivities for each cancer site.}

\item{start.time}{Numeric value specifying the start time of the simulation.}

\item{end.time}{Numeric value specifying the end time of the simulation.}

\item{obs.times}{Numeric vector specifying observation times.}

\item{start.states}{Numeric value specifying the initial states for each of the cancer types.}

\item{cancer_sites}{Character vector specifying the names of cancer sites.}
}
\value{
A data frame containing simulated observations for each cancer site.
}
\description{
This function simulates the progression of multiple cancer sites for a single individual over a specified time period.
}
\examples{
sim_multiple_cancer_indiv(ID = 1, cancer_sites = c("Lung", "Liver"),
                           rate_matrices = list(lungrate, liverate),
                           early_sensitivities = c(0.3, 0.2),
                           late_sensitivities = c(0.9, 0.8),
                           obs.times = c(0, seq(50, 75, by = 2)), end.time = 100,
                           start.time = 0,
                           start.state = 1)

}
